wip: exam template detail page
